reorganize extra divs into guide core
