+++ src/interface/guide/components/Preparation/FoodSubSection/FoodBuffPanels.tsx
@@ -1,8 +1,9 @@
 import Spell from 'common/SPELLS/Spell';
 import { QualitativePerformance } from 'parser/ui/QualitativePerformance';
-import CurrentFoodBuff from 'interface/guide/components/Preparation/FoodSubSection/CurrentFoodBuff';
-import { SideBySidePanels } from 'interface/guide/components/GuideDivs';
-import RecommendedFoodBuffs from 'interface/guide/components/Preparation/FoodSubSection/RecommendedFoodBuffs';
+
+import CurrentFoodBuff from './CurrentFoodBuff';
+import { SideBySidePanels } from '../../GuideDivs';
+import RecommendedFoodBuffs from './RecommendedFoodBuffs';
 
 interface Props {
   foodBuff?: Spell;
